+++ sentry-sidekiq/lib/sentry/sidekiq/sentry_context_middleware.rb
@@ -13,6 +13,20 @@ module Sentry
         span.set_data(Span::DataConventions::MESSAGING_MESSAGE_RECEIVE_LATENCY, latency) if latency
         span.set_data(Span::DataConventions::MESSAGING_MESSAGE_RETRY_COUNT, retry_count) if retry_count
       end
+
+      if ::Gem::Version.new(::Sidekiq::VERSION) >= ::Gem::Version.new("8.0.0")
+        def calculate_latency(job)
+          now_in_ms - job["enqueued_at"] if job["enqueued_at"]
+        end
+      else
+        def calculate_latency(job)
+          ((Time.now.to_f - job["enqueued_at"]) * 1000).round if job["enqueued_at"]
+        end
+      end
+
+      def now_in_ms
+        ::Process.clock_gettime(::Process::CLOCK_REALTIME, :millisecond)
+      end
     end
 
     class SentryContextServerMiddleware
@@ -40,7 +54,8 @@ module Sentry
         if transaction
           scope.set_span(transaction)
 
-          latency = ((Time.now.to_f - job["enqueued_at"]) * 1000).to_i if job["enqueued_at"]
+          latency = calculate_latency(job)
+
           set_span_data(
             transaction,
             id: job["jid"],