+++ sentry-sidekiq/spec/spec_helper.rb
@@ -11,6 +11,7 @@ require "sidekiq/cli"
 
 MIN_SIDEKIQ_6 = Gem::Version.new(Sidekiq::VERSION) >= Gem::Version.new("6.0")
 WITH_SIDEKIQ_7 = Gem::Version.new(Sidekiq::VERSION) >= Gem::Version.new("7.0")
+WITH_SIDEKIQ_8 = Gem::Version.new(Sidekiq::VERSION) >= Gem::Version.new("8.0.0")
 WITH_SIDEKIQ_6 = MIN_SIDEKIQ_6 && !WITH_SIDEKIQ_7
 
 require "sidekiq/embedded" if WITH_SIDEKIQ_7
@@ -54,6 +55,14 @@ RSpec.configure do |config|
 
   config.expect_with :rspec do |c|
     c.syntax = :expect
+  end
+
+  config.before :suite do
+    puts "\n"
+    puts "*" * 100
+    puts "Running with Sidekiq #{Sidekiq::VERSION}"
+    puts "*" * 100
+    puts "\n"
   end
 
   config.before :each do
@@ -273,6 +282,10 @@ def sidekiq_config(opts)
   WITH_SIDEKIQ_7 ? ::Sidekiq::Config.new(opts) : SidekiqConfigMock.new(opts)
 end
 
+def now_in_ms
+  ::Process.clock_gettime(::Process::CLOCK_REALTIME, :millisecond)
+end
+
 def execute_worker(processor, klass, **options)
   klass_options = klass.sidekiq_options_hash || {}
   # for Ruby < 2.6
@@ -283,12 +296,40 @@ def execute_worker(processor, klass, **options)
   jid = options.delete(:jid) || "123123"
   timecop_delay = options.delete(:timecop_delay)
 
-  msg = Sidekiq.dump_json(created_at: Time.now.to_f, enqueued_at: Time.now.to_f, jid: jid, class: klass, args: [], **options)
-  Timecop.freeze(timecop_delay) if timecop_delay
+  timestamps = if WITH_SIDEKIQ_8
+    current_time_ms = now_in_ms
+    {
+      created_at: current_time_ms,
+      enqueued_at: current_time_ms
+    }
+  else
+    {
+      created_at: Time.now.to_f,
+      enqueued_at: Time.now.to_f
+    }
+  end
+
+  msg = Sidekiq.dump_json(
+    jid: jid,
+    class: klass,
+    args: [],
+    **timestamps,
+    **options
+  )
+
+  if timecop_delay
+    Timecop.mock_process_clock = true
+    Timecop.freeze(timecop_delay)
+  end
+
   work = Sidekiq::BasicFetch::UnitOfWork.new('queue:default', msg)
+
   process_work(processor, work)
 ensure
-  Timecop.return if timecop_delay
+  if timecop_delay
+    Timecop.return
+    Timecop.mock_process_clock = false
+  end
 end
 
 def process_work(processor, work)